% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoAssignClientsToSurrogates}
\alias{AlgoAssignClientsToSurrogates}
\title{Map clients to surrogates
 @description assign equal (roughly) number of client nodes to each surrogate.
 @description Better assignment will be implemented in the future that takes into consideration number of requests (subscriptions in ICN) per node}
\usage{
AlgoAssignClientsToSurrogates(dns)
}
\arguments{
\item{dns}{map in the form of <d, n, p> where d is the dns node, n is the set of clients' nodes and p is the set of surrogates' nodes}
}
\value{
n,p array mapping n to p
}
\description{
Map clients to surrogates
 @description assign equal (roughly) number of client nodes to each surrogate.
 @description Better assignment will be implemented in the future that takes into consideration number of requests (subscriptions in ICN) per node
}
