% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoAssignSurrogateToOrigin}
\alias{AlgoAssignSurrogateToOrigin}
\title{Map Surrogate to Origin}
\usage{
AlgoAssignSurrogateToOrigin(D, O, E)
}
\arguments{
\item{D}{NxN distance matrix from each node to every other node.}

\item{O}{set of servers acting as origin servers}

\item{E}{set of edge servers}
}
\value{
a mapping of each e in E to some o in O
}
\description{
map surrogates to origins, that is the origin that the surrogate should contact to retrieve a copy of whatever items not cached

The map here is created by assigning to each surrogate the nearest origin
}
