% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoClosest}
\alias{AlgoClosest}
\title{Find k-nodes with the highest closeness in the graph}
\usage{
AlgoClosest(g, k, e = c())
}
\arguments{
\item{g}{graphNEL object}

\item{k}{k-closest nodes}

\item{e}{set of nodes to be excluded. primarly used to exclude CDN origins when selecting surrogates}
}
\value{
character vector of the node(s) name
}
\description{
Find k-nodes with the highest closeness in the graph
}
