% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoDij}
\alias{AlgoDij}
\title{Dijkstra weighted shortest tree(s) multicast algorithm}
\usage{
AlgoDij(g, eM, i)
}
\arguments{
\item{g}{network graph, used by dijkstra}

\item{eM}{edgeMatrix of the graph, providing edge: capacity, load, and hence weight}

\item{i}{Information item demand to be routed}
}
\value{
list of: d: distance matrix (publisher x subscribers) - used for debugging, pen: list of pens of each publisher - used for debugging, l: list of paths from each publisher to each subscriber - used for debugging, ts: multicast tress from the nearest publisher(s) to groups of subscribers, p: nearest publishers
}
\description{
The algorithm calculates a shortest path from each publisher to each subscriber. and acorrdingly group togheter (in a multicast tree) paths originating from the same publisher towards mutliple subscribers
}
