% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoDijAllPaths}
\alias{AlgoDijAllPaths}
\title{Calculate shortest paths}
\usage{
AlgoDijAllPaths(g, d)
}
\arguments{
\item{g}{network graph}

\item{d}{vector of edge weights}
}
\value{
N x N matrix of distances from each node to every other node. dataframe of all the SPR from each node to every other node
}
\description{
Calculate all shortest paths from each node to every other node in a graph g
}
