% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoDijDnsEdgeToOrigin}
\alias{AlgoDijDnsEdgeToOrigin}
\title{Route Surrogate to Origin, selected by DNS}
\usage{
AlgoDijDnsEdgeToOrigin(g, eM, i, ce, eo)
}
\arguments{
\item{g}{network graph}

\item{eM}{edge matrix of g}

\item{i}{item demand in the form of <i, s, p, ..>, although only one p is used for each client node in s}

\item{ce}{client-edge mapping, provided through DNS mapping in function DnsMapClientsToSurrogates}

\item{eo}{edge-origin mapping, providing edge server (surrogate) mapping to an origin - this comes through server selection policy, decoupled in IP from routing.}
}
\description{
Route a demand for item (non-live or non-streaming) according to conventional CDN based on selecting surrogates (or edge servers) that are nearest to the DNS

the algorithm calculates the shortest path from a client's access node (i.e. subscriber in the ICN world) to the surrogate selected through DNS mapping
}
\note{
there is only one surrogate that maps to each client, whereas the edge server has multiple choices but eventually will select one
}
