% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoDijNearestPop}
\alias{AlgoDijNearestPop}
\title{Find the nearest service point (publisher) to the subscriber}
\usage{
AlgoDijNearestPop(s, P, g)
}
\arguments{
\item{s}{character giving the subscriber}

\item{P}{character vector giving set of publisher}

\item{g}{network graph}

\item{eW}{numeric vector giving edge weights}
}
\value{
np nearest publisher
}
\description{
Find the nearest service point (publisher) to the subscriber
}
