% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoDijNeighbours}
\alias{AlgoDijNeighbours}
\title{Calculate shortest paths}
\usage{
AlgoDijNeighbours(g)
}
\arguments{
\item{g}{network graph}

\item{d}{vector of edge weights}
}
\description{
Calculate all shortest paths, through every neighbour of a node.
}
