% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoLargest}
\alias{AlgoLargest}
\title{Find k-nodes with the highest population in the graph}
\usage{
AlgoLargest(p, k, e = c())
}
\arguments{
\item{p}{a numeric vector of nodes' populations}

\item{k}{k-largest nodes}

\item{e(exclude)}{set of nodes to be excluded. primarly used to exclude CDN PoPs when selecting surrogates}
}
\value{
\code{n}: character vector of the node(s) name
}
\description{
Find k-nodes with the highest population in the graph
}
