% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoSelectFromCachedDijkstraRoutesForDnsMaps}
\alias{AlgoSelectFromCachedDijkstraRoutesForDnsMaps}
\title{Route demand on Dijkstra's cached routes}
\usage{
AlgoSelectFromCachedDijkstraRoutesForDnsMaps(i, D, ce, eo)
}
\arguments{
\item{i}{item demand in the form of <i, s, p, ..>, although only one p is used for each client node in s}

\item{D}{N x N matrix of the shortest distance between each pair of nodes, for all pairs}

\item{ce}{client-edge mapping, provided through DNS mapping in function DnsMapClientsToSurrogates}

\item{eo}{edge-origin mapping, providing edge server (surrogate) mapping to an origin - this comes through server selection policy, decoupled in IP from routing.}
}
\description{
Route a demand for item (non-live or non-streaming) according to conventional CDN based on selecting surrogates (or edge servers) that are nearest to the DNS

the algorithm calculates the shortest path from a client's access node (i.e. subscriber in the ICN world) to the surrogate selected through DNS mapping

Notice here there is only one surrogate that maps to each client, whereas the edge server has multiple choices but eventually will select one
}
