% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoSwing}
\alias{AlgoSwing}
\title{Select surrogates according to the 'swing' algorithm.
 @description The swing algorithm, orders the set of candidate surrogates according to their closeness to every other node. closeness is defined as the sum of all paths from the node to every other node
 @description It then selects the k surrogates by swinging between the nodes with shortest paths and nodes with longest paths, to ensure maximum coverage of the graph.}
\usage{
AlgoSwing(u, k = NULL, D, e)
}
\arguments{
\item{u}{vector giving per node population}

\item{k}{number of surrogates}

\item{D}{matrix of the distance from each node to every other node}
}
\value{
set of k or more surrogates
}
\description{
Select surrogates according to the 'swing' algorithm.
 @description The swing algorithm, orders the set of candidate surrogates according to their closeness to every other node. closeness is defined as the sum of all paths from the node to every other node
 @description It then selects the k surrogates by swinging between the nodes with shortest paths and nodes with longest paths, to ensure maximum coverage of the graph.
}
