% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{AlgoSwingCandidateConnector}
\alias{AlgoSwingCandidateConnector}
\title{Select candidate surrogates for each node. that would be the}
\usage{
AlgoSwingCandidateConnector(n, D, e = c())
}
\arguments{
\item{n}{nodes in the network}

\item{D}{shortest distance matrix}

\item{e}{nodes to be excluded from selection}
}
\description{
Select candidate surrogates for each node. that would be the
}
