% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{AssignAttrsToEdgeMatrix}
\alias{AssignAttrsToEdgeMatrix}
\title{Extend the edgeMatrix of a graph with a set of edge attributes}
\usage{
AssignAttrsToEdgeMatrix(eM, attrs = c("capacity", "weight", "pheromon",
  "rphermon"))
}
\arguments{
\item{eM}{edge matrix, possibly generated using GraphToEdgeMatrix(g)}

\item{attrs}{a character vector of the attributes to be added to the edge matrix (or the graph in that sense)}
}
\value{
updated edge matrix including added columns, one for each of the attributes of \code{attrs}
}
\description{
Extend the edgeMatrix of a graph with a set of edge attributes
}
