% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{Catalogue}
\alias{Catalogue}
\title{Construct content/service catalogue}
\usage{
Catalogue(N, d = "zipf", d.param = c(s = 1, ig = 0), v = 1, c = FALSE)
}
\arguments{
\item{N}{the number of ranks and when assuming 1 item per rank, N also means the number of items.}

\item{d}{character value, giving the probability distribution of the items. 'zipf' referes to zipf distribution.}

\item{v}{volume or capacity requirements of the item.}

\item{c}{create? <FALSE | TRUE>, default FALSE, decided whether or not to create the global ITEMS list}

\item{d.parm}{numeric vector giving the distribution parameters, e.g. for zipf: \code{s} exponent}
}
\value{
content catalogue in the form of a list of objects, each of which is <i, f, s, p, v>
}
\description{
Generate a list of item objects representing a content catalogue, each element is an item object in the form of <i, f, s, p, v>, where: i is the item identifier (also acts as the rank), f is the probability the item occurs in a dataset, s is set of subscribers and the number of their subscriptions, p is the set of publishers, v is the capacity requirement of the item.
}
