% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{CatalogueAppendPubSub}
\alias{CatalogueAppendPubSub}
\title{Append pub/sub to catalogue}
\usage{
CatalogueAppendPubSub(P, S, I, c = FALSE)
}
\arguments{
\item{P}{set of publishers (i.e. providers of information)}

\item{S}{set of subscribers (i.e. consumers)}

\item{I}{content catalogue}

\item{c}{update?<FALSE|TRUE>, default FALSE, decided whether or not to update the global ITEMS list}
}
\value{
updated content catalogue with publisher and/or subscriber information
update
remove list elements where there are no subscribers, as they are elements with no demand. it happens with some graphs and some light load
}
\description{
Update the content catalogue by adding publishers and/or subscribers information. This is equivelent to constructing the Information space of the RV
}
