% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{CatalogueAppendReflectors}
\alias{CatalogueAppendReflectors}
\title{Append reflectors to catalogue}
\usage{
CatalogueAppendReflectors(I, R, c)
}
\arguments{
\item{I}{content catalogue}

\item{R}{set of reflectors}

\item{c}{update? <FALSE | TRUE>, default FALSE, decided whether or not to update the global ITEMS list}
}
\value{
updated content catalogue to include reflectors
}
\description{
Assign reflectors to each streaming item in the content catalogue. This means the catalogue represents a collection of live streams, for each i in ITEMS, the names i is the STREAM IDENTIFIER.

Reflectors don't cache information locally, hence they are represneted by a standalone element in the 'i' list, not included in i$p or i$l
}
