% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{CatalogueSampleByVolume}
\alias{CatalogueSampleByVolume}
\title{Sample catalogue by volume}
\usage{
CatalogueSampleByVolume(I, v)
}
\arguments{
\item{I}{content catalogue}

\item{v}{total volume of all selected items}
}
\value{
Iv subset of I for which the total volume is v
}
\description{
select a subset of items, for which the total volume of all selected items, equals a defined value
}
