% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{Dns}
\alias{Dns}
\title{Assign DNS in the network}
\usage{
Dns(gu, D, k, opt)
}
\arguments{
\item{gu}{(g, u) list, where g is graphNEL of the  network and u is the population of each node}

\item{D}{NxN distance matrix}

\item{k}{number of DNS servers}

\item{opt}{numeric vector giving option of assignment algorithm. 1 == assign k largest population, 2 == assign k nodes with highest 'closeness' centrality}
}
\value{
list of nodes assinged as DNS points, whether local or public doesn't matter
}
\description{
Assign DNS in the network
}
