% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{DnsAssignClients}
\alias{DnsAssignClients}
\title{Assign clients to DNS points}
\usage{
DnsAssignClients(D, d, c = FALSE)
}
\arguments{
\item{D}{NxN matrix of distances from each node to every other node in the network graph}

\item{d}{vector of nodes acting as DNS}

\item{c}{boolean determining whether or not to create or update a globale variable for the DNS map}
}
\value{
<DNS, node> map associating each node with a DNS
}
\description{
assign network nodes to DNS nodes, this resembles assigning users to DNS servers
}
