% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{DnsAssignClientsToExistingMap}
\alias{DnsAssignClientsToExistingMap}
\title{Assing clients to existing DNS Map}
\usage{
DnsAssignClientsToExistingMap(dns, D, c = FALSE)
}
\arguments{
\item{dns}{incomplete DNS state, it maps DNS to surrogates but not yet to access nodes}

\item{D}{NxN matrix of distances from each node to every other node in the network graph}

\item{c}{boolean determining whether or not to create or update a globale variable for the DNS map}
}
\value{
<DNS, n, p> map associating each DNS with a set of nodes and a set of surrogates
}
\description{
assign access nodes to DNS existing but incomplete state (e.g. state only consists of DNS nodes and surrogates), this resembles assigning users to DNS servers
}
