% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{DnsAssignSurrogateToExistingMap}
\alias{DnsAssignSurrogateToExistingMap}
\title{Find surrogate (edge server) closest to local DNS (LDNS).}
\usage{
DnsAssignSurrogateToExistingMap(dns, p, D, c = FALSE)
}
\arguments{
\item{dns}{incomplete DNS map, having only DNS assignment to access nodes but surrogates not assigned to DNS.}

\item{p}{vector of nodes acting as edge servers (surrogates)}

\item{D}{NxN matrix of distances from each node to every other node in the network graph}

\item{c}{boolean determining whether or not to create a globale variable for the DNS map}
}
\value{
<DNS, n, p> map associating each DNS with a set of nodes and a set of surrogates
}
\description{
This resembels how the mapping system of traditional CDN performs global load balancing and local load balancing, by first selecting a CDN cluster then selecting servers within the cluster. the LDNS is then provided with identifiers for those servers.

See the paper "End-User Mapping: Next Generation Request Routing for Content Delivery" for a more comprehensive description of how CDN mapping works
}
