% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{DnsMap}
\alias{DnsMap}
\title{Construct DNS map}
\usage{
DnsMap(gu, D, k, opt, p, c = FALSE)
}
\arguments{
\item{D}{NxN matrix of distance from each node to every other node in the graph. if missing calculate inside the function}

\item{k}{number of DNS servers}

\item{opt}{numeric vector giving option of assignment algorithm. 1 == assign k largest population, 2 == assign k nodes with highest 'closeness' centrality}

\item{p}{vector of nodes acting as edge servers (surrogates)}

\item{c}{boolean determining whether or not to create a globale variable for the DNS map}
}
\value{
list mapping DNS nodes to client and surrogate nodes, and an array directly mapping clients to surrogates/origins
}
\description{
Generate DNS map, start by selecting DNS nodes, then assign client's nodes to DNS, assing surrogates and origins to DNS node, and finally map client's to surrogates and origins.
}
