% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{DnsMapClientsAndUpdateSubscribersList}
\alias{DnsMapClientsAndUpdateSubscribersList}
\title{Map clients to DNS and update the subscribers state}
\usage{
DnsMapClientsAndUpdateSubscribersList(d, S, D, c = TRUE)
}
\arguments{
\item{d}{vector of nodes acting as DNS}

\item{S}{subscribers state in the form <s, v, i>}

\item{D}{NxN matrix of distances from each node to every other node in the network graph}

\item{c}{boolean determining whether or not to update the globale list SUBSCRIBERS in the global enviroment}
}
\value{
updated subscribers state in the form <s,v, i, dns>
}
\description{
Update subscribers state - i.e. the subscribers list created in subs.R, which includes the list of items requested by end uesers, this resembles assigning users to DNS servers
}
