% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{DnsMapClientsToSurrogates}
\alias{DnsMapClientsToSurrogates}
\title{Map clients (subscribers) to surrogates}
\usage{
DnsMapClientsToSurrogates(dns)
}
\arguments{
\item{dns}{list mapping client nodes (equivelent to subscribers in the ICN domain) to a DNS node and mapping surrogates to DNS}
}
\value{
a map of client nodes to surrogate nodes. this is equivelent to the function of the mapping system by conventional CDN employing DNS redirects with local and global balancing
}
\description{
Map clients to surrogates in the DNS map, when there are multiple surrogates associated with a DNS,

Perform a simple round robin algorithm to load balance the surrogates.
}
