% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ExtractPath}
\alias{ExtractPath}
\title{modified version of RBGL extractPath function}
\usage{
ExtractPath(s, f, pens)
}
\arguments{
\item{s}{source of the path}

\item{f}{destination of the path}

\item{pens}{pens of the source s}
}
\value{
vector of nodes giving the path from s to f
}
\description{
to include wavelength augumented graphs
}
