% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenActiveUsers}
\alias{GenActiveUsers}
\title{Generate active users}
\usage{
GenActiveUsers(gun, scale, tests, path)
}
\arguments{
\item{gun}{network graphNEL}

\item{scale}{0 > vector > 1.0 of the ratio of demands to capacity in the form of \code{(lb, ub, st)}, lb/ub: lower/upper bound, st: step}

\item{tests}{number of random permutations}

\item{path}{the directory path for where to save the demands
print(nodes.population)
generate lognormal that is in the size of total.demand}
}
\description{
Generate scaling vectors of demand volumes, each scale has multiple random generated volumes in the count of 'tests'
}
