% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenCataloguePermutation}
\alias{GenCataloguePermutation}
\title{Generate catalogue permutations}
\usage{
GenCataloguePermutation(.N, .d, .d.param, .v, .path)
}
\arguments{
\item{.N}{Size of the content catalogue, i.e. number of items in the catalogue}

\item{.d}{popularity distribution of the items (e.g. Zipf, lognormal,..), currently only Zipf is supported}

\item{.d.param}{distribution parameters, these are the factors that impact distribution shape. e.g. Zipf exponent and the ignore ratio (i.e. how much of the top level to ignore) (s = 1, ig = 0)}

\item{.v}{vector of double values, defining the possible volume per item}

\item{.path}{string giving the directory path for where to store the permutations}
}
\description{
generate (and save) permutations of content catalogues for multiple tests and possibly multiple load values.

saved permutations can be loaded at the start of a simulation rather than be calculated then, to save simulation time
}
