% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenDnsAndSurrogateToOriginMap}
\alias{GenDnsAndSurrogateToOriginMap}
\title{Generate DNS and surrogate-to-origins permutations}
\usage{
GenDnsAndSurrogateToOriginMap(opt, k, O, E, P, gund, tests, dns.path, eo.path)
}
\arguments{
\item{opt}{DNS selection policy}

\item{k}{number of LDNS points}

\item{O}{names and configurations of origins permutation, length of the list equals tests}

\item{E}{names and configurations of of surrogates permutation, length of the list equals tests}

\item{P}{names of publishers (i.e. joined O and E) permutation, length of list equals tests}

\item{tests}{number of permutations}

\item{dns.path}{string giving the path to write the permutations of the DNS maps}

\item{eo.path}{string giving the path to write the permutations of the surrogate-to-origin maps}

\item{gun}{network graph, population per node and graph name - i.e <g, u, g.name>}

\item{D}{NxN matrix of distance from each node to every other node}
}
\description{
generate DNS and surrogate-to-origin mappings for permutations of origins and surrogates for a single set of configurations
}
\note{
, the function is intermidiary, as it should normally be called from within \code{GenDnsAndSurrogateToOriginMapPermutations}
}
