% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenDnsAndSurrogateToOriginMapPermutations}
\alias{GenDnsAndSurrogateToOriginMapPermutations}
\title{Generate DNS map permutations}
\usage{
GenDnsAndSurrogateToOriginMapPermutations(cfg, P, gund, tests, path)
}
\arguments{
\item{cfg}{DNS configurations, currently including: \code{k} number of LDNS points and \code{opt} DNS selection policy}

\item{P}{list of origins (O), surrogates(E) and their merge to publishers (P) permutations elements, provided by the function \code{GenPublishersPermutations}}

\item{gund}{network graph, population per node, graph name and distance matrix - i.e <g, u, g.name, D>}

\item{tests}{number of permutations}

\item{path}{string giving the paths to write the permutations of the DNS maps and surrogate-to-origin maps
calculate file offset
create DNS and surrogate-to-origin mapping for every combination of origins and surrogates}

\item{D}{NxN matrix of distance from each node to every other node}
}
\description{
Generate DNS map of client nodes to surrogate and origin nodes, also generate a surrogate to origin map
}
