% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenDnsMap}
\alias{GenDnsMap}
\title{Generate DNS map}
\usage{
GenDnsMap(P, opt, k, D, gun, tests, dns.path)
}
\arguments{
\item{P}{names of publishers (i.e. joined O and E) permutation, length of list equals tests}

\item{opt}{DNS selection policy}

\item{k}{number of LDNS points}

\item{D}{NxN matrix of distance from each node to every other node}

\item{gun}{network graph, population per node and graph name - i.e <g, u, g.name>}

\item{tests}{number of permutations}

\item{dns.path}{string giving the path to write the permutations of the DNS maps}
}
\description{
generate DNS map for permutations of origins and surrogates for a single set of configurations
}
