% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenEoMap}
\alias{GenEoMap}
\title{Generate surrogate-to-origins map}
\usage{
GenEoMap(O, E, D, tests, eo.path)
}
\arguments{
\item{O}{names and configurations of origins permutation, length of the list equals tests}

\item{E}{names and configurations of of surrogates permutation, length of the list equals tests}

\item{D}{NxN matrix of distance from each node to every other node}

\item{tests}{number of permutations}

\item{eo.path}{string giving the path to write the permutations of the surrogate-to-origin maps}
}
\description{
Generate surrogate-to-origin maps for permutations of origins and surrogates for a single set of configurations
}
