% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenMergedOriginsAndSurrogates}
\alias{GenMergedOriginsAndSurrogates}
\title{Generate merged list of origins and permutation}
\usage{
GenMergedOriginsAndSurrogates(O, E, .gname, .tests, .path)
}
\arguments{
\item{O}{list of origins}

\item{E}{list of surrogates}

\item{.gname}{network graph name}

\item{.tests}{numeric value giving number of permutations inside each list}

\item{.path}{string giving the paths to write the publishers permutations}
}
\description{
merge the list of origins with the list of surrogates in the one list of publishers
}
