% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenNexpPmf}
\alias{GenNexpPmf}
\title{Generate an Exponential distribution}
\usage{
GenNexpPmf(N, ig = 0)
}
\arguments{
\item{N}{the number of ranks when assuming 1 item per rank}

\item{ig}{ignore ratio (0.0, 0.5): fraction of the top of the distribution to ignore}

\item{s}{the distribution exponent}
}
\value{
name/value numeric vector of zipf-based pmf, whereby name is the rank of an element and value is the fraction of time it appears in a dataset
}
\description{
The generated value is the probablity of occurance in 100 events.
}
