% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenOriginsPermutations}
\alias{GenOriginsPermutations}
\title{Generate origins permutations}
\usage{
GenOriginsPermutations(.opt, .k, .a, .cc, .gund, .tests, .path)
}
\arguments{
\item{.opt}{numeric value indicating the selection policy, this corresponds to the service placement algorithm. currently there are 3 greedy algorithms: largest population first, highest closeness first and swing}

\item{.k}{numeric value giving #origins}

\item{.a}{numberic value giving min storage size}

\item{.cc}{content catalogue instance. A list that includes <I, Z, N, d, d.params> whereby: I content objects, Z popularity distribution, N number of items, d name of popularity distribution, d.params parameters used to generate the distribution}

\item{.gund}{network graph, population per node, graph name and distance matrix - i.e <g, u, g.name, D>}

\item{.tests}{number of permutations}

\item{.path}{string giving the full paths to write the origins permutations, including the object name}
}
\description{
Generate (and save) origins permutations, those are nodes holding copies of every item in the content catalogue, there are no restriction on their storage capacity

The saved permutations can be loaded at the start of a simulation rather than be calculated then, to save simulation time
}
