% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenPublishersPermutations}
\alias{GenPublishersPermutations}
\title{Generate publishers' permutations}
\usage{
GenPublishersPermutations(cc, ocfg, ecfg, gund, tests, path)
}
\arguments{
\item{cc}{content catalogue in the form of a list that includes: the content catalogue items (I) and the popularity distribution of I (Z)}

\item{ocfg}{origin configurations}

\item{ecfg}{surrogate configurations}

\item{gund}{network graph, population per node, graph name and distance matrix - i.e <g, u, g.name, D>}

\item{tests}{number of permutations}

\item{path}{string giving the paths to write the publishers permutations}
}
\value{
list of origins and surrogates permutations, as well as the merge of both to give 'publishers' sets. All have been saved
}
\description{
Generate (and save) origins and surrogates - publishers - permutations for a single content catalogue, origins are generated first therefore no nodes need to be excluded

Then for each origin instance, permutations of surrogates are generated according to surrogates configurations which excludes the origins nodes.
}
