% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenPubSubState}
\alias{GenPubSubState}
\title{Generate pub/sub state}
\usage{
GenPubSubState(.P, .cc, .S, .scale, .tests, .path)
}
\arguments{
\item{.P}{permutations of publishers generated for .cc for an snapshot of configurations, the length of the list should equal tests}

\item{.cc}{permutations of content catalogue}

\item{.S}{permutations of subscribers generated for .cc, the length of the list should equal \code{tests} x \code{#steps in scale}}

\item{.scale}{load of subscriptions}

\item{.tests}{number of permutations}

\item{.path}{path to store the generated data
mapply is the fastes approach from the 3 options: 1- serial for , 2- parallel foreach and 3- mapply}
}
\description{
generate (and save) pub/sub state according to saved permutations of clients (S), origins (O) and surrogates (E)

saved state can be loaded at the start of a simulation rather than be calculated then.
}
