% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenRndId}
\alias{GenRndId}
\title{simulate multiple instances}
\usage{
GenRndId(x)
}
\arguments{
\item{x}{number of ids to generate}
}
\value{
unique random string, consisting of letters and numbers
}
\description{
Generate random unique identifier, used to link different objects of the same simulation to each other.

The function is taken from the R-blog: https://www.r-bloggers.com/generating-unique-random-ids/
}
