% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenSimData}
\alias{GenSimData}
\title{Generation simulation data}
\usage{
GenSimData(cfgs, config.active = "default")
}
\arguments{
\item{cfgs}{parsed configuration groups from a YAML file, see function \code{ParseConfig}}

\item{config.active}{configuration enviroment of the logger
set logging level
check stored data, if not generate then store. start first with input catalogue
retreive users permutations depending on scale of load. generate if not already stored.
check if subscribers stored, if not generate and store}
}
\description{
Generate (and save) simulation data provided a simulation configuration file
}
\details{
the configs should include all the necessary groups required to generate the input data. Generated data will be stored with consisten name, whereby the catalogue cfgs will make up the name of the catalogue subdirectory. inside each catalogue directory, there will be the permutations generated for the catalgoue. The name of each data file follows the format \code{ObjectName-graph_params-object_Params.RData}
}
