% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenSubsPermutations}
\alias{GenSubsPermutations}
\title{Generate subscribers' permutations}
\usage{
GenSubsPermutations(.I, .gun, .cfgs, .users, .path)
}
\arguments{
\item{.I}{a sorted set, representing the global set of items in the network. I is the probabilities of occurance of the items}

\item{.gun}{network graph, population per node and graph name - i.e <g, u, g.name>}

\item{.cfgs}{load configs, must include: \code{load} for scale of load on the network and \code{tests} number of random permutations.}

\item{.users}{list of active users per node, for each load value in \code{scale} and each test in \code{tests}. If this users is missing, the function will try to load from stored list or calculate the list if not stored using \code{\link{GenActiveUsers}}}

\item{.path}{the full path (including filename) for where to save the demands}
}
\description{
Generate randomised and scaling subscriptions for multiple tests and possibly multiple load values
}
