% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenZipfPmf}
\alias{GenZipfPmf}
\title{Generate a Zipf distribution}
\usage{
GenZipfPmf(s = 1, N, ig = 0)
}
\arguments{
\item{s}{the distribution exponent}

\item{N}{the number of ranks when assuming 1 item per rank}

\item{ig}{ignore ratio (0.0, 0.5): fraction of the top of the distribution to ignore}
}
\value{
name/value numeric vector of zipf-based pmf, whereby name is the rank of an element and value is the fraction of time it appears in a dataset
}
\description{
The generated value is the probablity of occurance in 100 events.
}
