% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetActiveUsers}
\alias{GetActiveUsers}
\title{Get active users in network}
\usage{
GetActiveUsers(.gun, .cfgs, .path)
}
\arguments{
\item{.gun}{network graph, population per node and graph name - i.e <g, u, g.name>}

\item{.cfgs}{load configs, must include: \code{load} for scale of load on the network and \code{tests} number of random permutations.}

\item{.path}{the directory path for where to save the demands}
}
\value{
list of active users
}
\description{
Get the list of active users per node in a graph. if not available then calculate it
}
