% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetCatalogue}
\alias{GetCatalogue}
\title{Get Catalogue}
\usage{
GetCatalogue(N, d, d.param, v, path)
}
\arguments{
\item{N}{size of the catalogue}

\item{d}{popularity distribution}

\item{d.param}{parameters of the popularity distribution}

\item{v}{volume range of the items}

\item{path}{to where the catalogue directory and the catalogue object should be stored}
}
\description{
Get content catalogue, i.e. check if catalogue is stored and if not, then generate and store in a new directory of name made up from the catalogue config. params
}
