% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetCataloguePath}
\alias{GetCataloguePath}
\title{get the directory path to the content catalogue}
\usage{
GetCataloguePath(N, d, dparams, v, dir = "")
}
\arguments{
\item{dir}{parent directory of the catalogue}

\item{cfgs}{list of config variables for a single catalogue}
}
\value{
path to the catalogue directory
}
\description{
get the directory path to the content catalogue
}
