% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetComplementaryItems}
\alias{GetComplementaryItems}
\title{Get complementary items}
\usage{
GetComplementaryItems(J, I)
}
\arguments{
\item{J}{subset of content items that belongs to I}

\item{I}{the content catalogue}
}
\value{
vector of the complementary items
}
\description{
get complementary items in the content catalogue, i.e. items in the catalogue that does not appear in a subset
}
