% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetDnsAndSurrogateToOriginMaps}
\alias{GetDnsAndSurrogateToOriginMaps}
\title{Get DNS based maps}
\usage{
GetDnsAndSurrogateToOriginMaps(cc.dir, cc = NULL, P, gund, cfgs)
}
\arguments{
\item{cc.dir}{subdirectory of the content catalogue, name of the dir contain all the configurations (metadata) defining the catalogue}

\item{cc}{the catalogue in the list of catalogues. The list is loaded into a separate ArgsEnv enviroment. default to NULL to indicate no number is selected}

\item{P}{list of publishers permutations, including \code{O} origins, \code{E} edges (surrogates) and \code{P} the merge of both}

\item{gund}{network graph, population per node graph name and Distance matrix - i.e <g, u, g.name, D>}

\item{cfgs}{common configs, as parsed by \code{ParseConfig}
rotate over surrogates, first move to the chunck of rows of particular origins expr (combination). \code{j -1} is to avoid special cases where \code{j == (dlen * elen)}, which would result in \code{k == 0}}
}
\description{
Get DNS based maps, including: DNS-based client-to-surrogate and surrogate-to-origin mappings

This resembels DNS redirect in native IP-based CDN
}
\note{
Edge to Origin mapping is not bound by the number of LDNSes, because CDN management handle that independently from LDNSes
}
