% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetDnsToSurrogateArrayMap}
\alias{GetDnsToSurrogateArrayMap}
\title{convert DNS-to-surrogate map format from list to array}
\usage{
GetDnsToSurrogateArrayMap(d, p)
}
\arguments{
\item{d}{name of DNS node}

\item{p}{set of surrogates for a DNS node}
}
\value{
array form of the dns list
}
\description{
convert DNS-to-surrogate map format from list to array
}
