% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetEdgesOnMulticastTree}
\alias{GetEdgesOnMulticastTree}
\title{Extract the set of edges in a tree list}
\usage{
GetEdgesOnMulticastTree(t)
}
\arguments{
\item{t}{tree in the form of list, each element represent a branch in the tree}
}
\value{
matrix of edges in the tree
}
\description{
common edges will be merged into one edge indicating multicast
}
