% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetEdgesOnPath}
\alias{GetEdgesOnPath}
\title{Convert path to set of edges}
\usage{
GetEdgesOnPath(p)
}
\arguments{
\item{p}{path as a vector of nodes}
}
\value{
matrix of edges in the form <from, to>
}
\description{
Convert path to set of edges
}
