% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetGraph}
\alias{GetGraph}
\title{Get network graph(s)}
\usage{
GetGraph(cfg, paths)
}
\arguments{
\item{cfg}{graph configurations, including: gdir where the graphs are stored, gset: index of graphs and augument, whether or not to augument the graph}

\item{paths}{string vector of at least 2 elements, first is paths['home'], indicating home directory and second paths['store'] indicating storage (write) directory}
}
\value{
list of gun <g, u, n>, i.e. graph object, population per node, graph name
}
\description{
Get network graph(s)
}
