% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetGraphName}
\alias{GetGraphName}
\title{Get graph name for list of graphs}
\usage{
GetGraphName(gset)
}
\arguments{
\item{gset}{list of graph objects, each of which includes element \code{g.name}}
}
\value{
graph name as stated by $g.name
}
\description{
Get graph name for list of graphs
}
