% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetIcnMcCapacity}
\alias{GetIcnMcCapacity}
\title{Get ICN multicst capacity results
Get ICN provisioned capacity when co-incidental multicast is in place, given a catchment interval \code{t} seconds and a total running time of \code{L} seconds}
\usage{
GetIcnMcCapacity(cc, o, g, t, L)
}
\arguments{
\item{cc}{pub/sub state for a content catalogue, required as it gives the number of subscriptions per item}

\item{o}{list object of the multicast tree - must include: trees: i.e. list of paths in the multicast trees, added: list of added values in unicast paths, this is merely a container so don't have to recreat it, and total: again a container for total values.}

\item{g}{network graph list including graph, population per node, name, distance matrix and edge matrix}

\item{t}{catchment interval in seconds}

\item{L}{total running period in seconds}
}
\description{
Get ICN multicst capacity results
Get ICN provisioned capacity when co-incidental multicast is in place, given a catchment interval \code{t} seconds and a total running time of \code{L} seconds
}
