% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetIcnMcCapacityForOneDemand}
\alias{GetIcnMcCapacityForOneDemand}
\title{update multicast load on each tree}
\usage{
GetIcnMcCapacityForOneDemand(iL, o, t, L)
}
\description{
update multicast load on each tree
}
