% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetIcnUnicastEmResults}
\alias{GetIcnUnicastEmResults}
\title{load Edge Matrix (eM) for the ICN results data of unicast}
\usage{
GetIcnUnicastEmResults(filename)
}
\arguments{
\item{filename}{of the result data}
}
\description{
load Edge Matrix (eM) for the ICN results data of unicast
}
