% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetIcnUnicastPathLengthEcdfResults}
\alias{GetIcnUnicastPathLengthEcdfResults}
\title{load the paths of the ICN simulation results for unicast}
\usage{
GetIcnUnicastPathLengthEcdfResults(filename)
}
\arguments{
\item{filename}{of the results data}

\item{D}{NxN matrix of all the routes}
}
\description{
load the paths of the ICN simulation results for unicast
}
