% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetIcnUnicastPathLengthResults}
\alias{GetIcnUnicastPathLengthResults}
\title{load the paths of the ICN simulation results for unicast}
\usage{
GetIcnUnicastPathLengthResults(filename)
}
\arguments{
\item{filename}{of the results data}
}
\description{
load the paths of the ICN simulation results for unicast
}
