% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetIpEmResults}
\alias{GetIpEmResults}
\title{load Edge Matrix (eM) for the IP results data}
\usage{
GetIpEmResults(filename)
}
\arguments{
\item{filename}{of the result data}
}
\description{
load Edge Matrix (eM) for the IP results data
}
