% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetIpPathLengthEcdfResults}
\alias{GetIpPathLengthEcdfResults}
\title{load the tress of the IP simulation results}
\usage{
GetIpPathLengthEcdfResults(filename)
}
\arguments{
\item{filename}{of the results data}
}
\description{
load the tress of the IP simulation results
}
