% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetIpPathLengthResults}
\alias{GetIpPathLengthResults}
\title{load the tress of the IP simulation results}
\usage{
GetIpPathLengthResults(filename)
}
\arguments{
\item{filename}{of the results data}
}
\description{
load the tress of the IP simulation results
}
