% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetMatchingRowIndex}
\alias{GetMatchingRowIndex}
\title{get matching row indices between an Edge Matrix and a sub matrix of a path}
\usage{
GetMatchingRowIndex(eM, es)
}
\arguments{
\item{eM}{edge matrix}

\item{es}{subset of edges in eM}
}
\description{
get matching row indices between an Edge Matrix and a sub matrix of a path
}
