% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetNamesAndCfgs}
\alias{GetNamesAndCfgs}
\title{Get publishers' names and configs}
\usage{
GetNamesAndCfgs(O, tests)
}
\arguments{
\item{O}{nested list of origins, each list in the list corresponds to a set of configuration params}

\item{tests}{the number of O permutations}
}
\value{
nested list of origins names, in addition to origins cfg
}
\description{
get origins, surrogates, publishers (both origins and surrogates), DNS or subscribers names and configs, provided a list of origins.

The list structure is typically in the form O[[combo.row.id]][[test.num]], O$cfg
}
