% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetOrigins}
\alias{GetOrigins}
\title{Get origins}
\usage{
GetOrigins(opt, k, a, gund, tests, path)
}
\arguments{
\item{opt}{numeric value indicating the selection policy, this corresponds to the service placement algorithm. currently there are 3 greedy algorithms: largest population first, highest closeness first and swing}

\item{k}{number of origins}

\item{a}{minimum storage unit in GB}

\item{gund}{network graph, population per node and graph name - i.e <g, u, g.name, DM, AM>, DM/AM: Distance/Adjacency Matrix}

\item{tests}{number of random permutations}

\item{path}{full path to retreive or write the origins list}
}
\description{
Retrieve the stored permutations of Origins. if not stored, then calculate, store and return it.
}
\note{
NOT COMPLETE YET!!
}
