% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetPublishersPermutations}
\alias{GetPublishersPermutations}
\title{Get publishers permutations}
\usage{
GetPublishersPermutations(cc, gund, cfgs)
}
\arguments{
\item{cc}{content catalogue, including \code{dp} the directory path to the subdirectory of the content catalogue}

\item{gund}{network graph, population per node, graph name and Distance matrix - i.e <g, u, g.name, D>}

\item{cfgs}{various common cfgs, must include: \code{ocfg} for origin configs, \code{ecfg} edge (surrogate) configs, and \code{tcfg} - including \code{tcfg$tests} - for load configs.
set up output containser
-------------------------
verify origins first
get/generate surrogates....
get/generate publishers (i.e. merge of origins and surrogates)
load already-stored publishers files}
}
\description{
Get publishers permutations if stored, if not generate then store
}
