% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetPubSubState}
\alias{GetPubSubState}
\title{get Pub/Sub (i.e. PCF) state for a catalogue(s)}
\usage{
GetPubSubState(cc.dir, cc = NULL, P, S, gund, cfgs)
}
\arguments{
\item{cc.dir}{subdirectory of the content catalogue, name of the dir contain all the configurations (metadata) defining the catalogue}

\item{cc}{the catalogue in the list of catalogues. The list is loaded into a separate ArgsEnv enviroment. default to NULL to indicate no number is selected}

\item{P}{list of publishers permutations, including \code{O} origins, \code{E} edges (surrogates) and \code{P} the merge of both}

\item{S}{list of subscription permutations}

\item{cfgs}{common configs, as parsed by \code{ParseConfig}, must include: \code{paths} to where to place/retrieve data, \code{cores} number of cpus to use in running, \code{eoexpr} combinations of regular expressions used in naming files and \code{tcfg} for load configs}
}
\value{
pub/sub state of the catalogue, i.e. added publishers and subscribers to the catalogue object
check which files are missing, generate and store them.
}
\description{
get Pub/Sub (i.e. PCF) state for a catalogue(s)
}
