% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetStorageResults}
\alias{GetStorageResults}
\title{get the pub/sub state to obtain the storage capacity}
\usage{
GetStorageResults(filename)
}
\arguments{
\item{filename}{of the pub/sub state}

\item{scale}{load on the network, this is the outer index of the nest list of pub/subs}

\item{tests}{number of simulations, that is the inner scale of the list}
}
\description{
get the pub/sub state to obtain the storage capacity
}
