% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetStoredObjects}
\alias{GetStoredObjects}
\title{get list of stored objects}
\usage{
GetStoredObjects(filenames)
}
\arguments{
\item{filenames}{}
}
\value{
loaded objects
}
\description{
get list of stored objects
}
