% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetSubscribersPermutations}
\alias{GetSubscribersPermutations}
\title{Get subscribers permutations}
\usage{
GetSubscribersPermutations(cc, gun, users, cfgs, path)
}
\arguments{
\item{cc}{content catalogue, including \code{dp} the directory path to the subdirectory of the content catalogue}

\item{gun}{network graph, population per node and graph name - i.e <g, u, g.name>}

\item{users}{numeric vector of active users per node}

\item{cfgs}{load configs, must include: \code{load} for scale of load on the network and \code{tests} number of random permutations.}

\item{path}{character vector, gives the full path (including filename) to retreive or store and retreive the object
verify subscribers exisit}
}
\description{
Get subscribers permutations if stored, if not then generate and store
}
