% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphAugumentAccessNodes}
\alias{GraphAugumentAccessNodes}
\title{Augument the graph to add access node to each node with capacity between the original node and the access node = total edge capacity of the original node.
This means the access node cannot push to the origin node any more traffic that the sum of capacities on all of its edges.}
\usage{
GraphAugumentAccessNodes(g, eM)
}
\arguments{
\item{g}{core network graph}

\item{eM}{core edge Matrix of the graph with capacities}
}
\value{
augumented graph including core + access nodes

augumented edge Matrix of core + access edges
}
\description{
Augument the graph to add access node to each node with capacity between the original node and the access node = total edge capacity of the original node.
This means the access node cannot push to the origin node any more traffic that the sum of capacities on all of its edges.
}
