% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphAugumentTreeAccessNetwork}
\alias{GraphAugumentTreeAccessNetwork}
\title{Augument Access network in the form of a k-ary tree, with {k,h} parameters.}
\usage{
GraphAugumentTreeAccessNetwork(gu, k, m)
}
\arguments{
\item{gu}{\code{g} graph object of the core network, to be augumented. \code{u} a numeric vector of the number of active users per each core node.}

\item{k}{tree-arity, i.e. the maximum number of children each node can have}

\item{m}{numeric value defining the number of users per access node, u/m gives the total number of leafs}
}
\value{
augument graph with access and core network nodes
}
\description{
Augument Access network in the form of a k-ary tree, with {k,h} parameters.
}
