% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphAugumentTreeAccessZoo}
\alias{GraphAugumentTreeAccessZoo}
\title{Augument access network as a balanced tree}
\usage{
GraphAugumentTreeAccessZoo(read.path, write.path, augument.params = c(k = 20,
  m = 1000))
}
\arguments{
\item{read.path}{path to read list of zoo graph object(s) of core node(s). This can be the imported zoo graphs using GraphImportZoo}

\item{write.path}{path to write the R list of all augumented graphs. filename is azoo.RData}

\item{augument.params}{\code{k} the tree-arity parameter defining the number of childer per parent. \code{m} maximum number of users that can be handled by a leaf node}
}
\value{
\code{files} list of file names of zoo topologies that have been successfuly augumented.

\code{graphs} list of \code{g} the core/access augumented graph. \code{u} the number of active users of every node in the augument graph, core and aggregation nodes have \code{0} population as they don't have users attached to them. \code{access} list of graph objects of all the access trees attached to the core graph.
}
\description{
Augument a core network topology that has population associated with it into access/core network.
The augumentation takes the population associated with a core node and use it to create a k-ary tree access network
other design parameters that define the tree is \code{k} the arity parameter (i.e. the max number of children for each parent in the tree), and
\code{m} the max number of users to be associated with any leaf node.
}
\seealso{
\code{\link{GraphLoadZoo}}, \code{\link{GraphImportZoo}}
}
\author{
Mays AL-Naday
}
