% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphCreateTreeNetwork}
\alias{GraphCreateTreeNetwork}
\title{Create a k-ary tree access network for a set of active users}
\usage{
GraphCreateTreeNetwork(n, v, c, k)
}
\arguments{
\item{n}{the number of nodes in the graph. used as a starting index for numbering the nodes in the access network to avoid node id clash with the core graph}

\item{v}{a (name,value) number where name is the node id of a core node, and value is the number of active users associated with the node}

\item{c}{design parameter of the maximum number of users that can be served by an access (leaf) node}

\item{k}{tree-arity design parameter, i.e. maximum number of childer that can be aggregated by one parent in the tree}
}
\value{
R list of: \code{h}: the hight of the tree, \code{parents}: the number of parents at each level of the tree, \code{edges} edge matrix of the tree graph, \code{nodes} vector of nodes in the tree graph, \code{u} vector of populations per node in the tree, \code{sum(u) == v}. Notice that the core node and parents (i.e. aggregate nodes) have a population of \code{zero}.
}
\description{
Create a k-ary tree access network for a set of active users
}
