% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphImportZoo}
\alias{GraphImportZoo}
\title{Import Zoo topologies}
\usage{
GraphImportZoo(read.path, write.path, ext = "graphml")
}
\arguments{
\item{read.path}{path to read graph objects in either gml or graphml format}

\item{write.path}{path to write the R list of all read graphs:node_population object}

\item{extension}{graph format (zoo provides either 'gml' or 'graphml')}
}
\value{
list of \code{{g, u}} objects of the zoo list of topology files. \code{g} graphNEL, \code{u} nodes population
}
\description{
Import Zoo topology graphs and extrapolate node population of each node in every graph, based on the WorldPopulation
}
\seealso{
\code{\link{GraphLoadZoo}}
}
\author{
Mays AL-Naday
}
