% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphLoadZoo}
\alias{GraphLoadZoo}
\title{Load Zoo graphs}
\usage{
GraphLoadZoo(read.path, write.path, ext = "graphml")
}
\arguments{
\item{read.path}{path to read graph objects in either gml or graphml format}

\item{write.path}{path to write the R list of all read graphs:node_population object}

\item{extension}{graph format (zoo provides either 'gml' or 'graphml')}
}
\value{
list of \code{{g, u}} objects of the zoo list of topology files. \code{g} graphNEL, \code{u} nodes population
}
\description{
Load list of zoo graphs, if stored. otherwise, read, store then load
}
\seealso{
\code{\link{GraphImportZoo}}
}
\author{
Mays AL-Naday
}
