% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphSetEdgeAttr}
\alias{GraphSetEdgeAttr}
\title{Set the edge attribute of graph to value}
\usage{
GraphSetEdgeAttr(eM, val = 1, attr)
}
\arguments{
\item{eM}{edge matrix containing graph attribute to be set}

\item{val}{value to set the edge attribute, accepts a vector but cannot be NULL, NA or NaN}
}
\value{
Edge matrix with attribue set to val
}
\description{
Set the edge attribute of graph to value
}
