% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphSetLoadOnAdjMatrix}
\alias{GraphSetLoadOnAdjMatrix}
\title{Set the edge load in a NxN adjacency matrix}
\usage{
GraphSetLoadOnAdjMatrix(eM, val = 1, attr)
}
\arguments{
\item{eM}{adjaceny matrix}

\item{val}{value to set the edge load, accepts a vector but cannot be NULL, NA or NaN}
}
\value{
adjacency matrix with load set to val
}
\description{
Set the edge load in a NxN adjacency matrix
}
