% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphToAdjMatrix}
\alias{GraphToAdjMatrix}
\title{Convert graph to N x N adjacency matrix where nxm cell gives the load on edge n->m}
\usage{
GraphToAdjMatrix(g)
}
\arguments{
\item{g}{network graph}

\item{attrs}{vector of strings, listing the edge attributies}
}
\value{
edge matrix with added columns for edge attributes
}
\description{
Convert graph to N x N adjacency matrix where nxm cell gives the load on edge n->m
}
