% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphToEdgeMatrix}
\alias{GraphToEdgeMatrix}
\title{Convert a graphNEL to an edge matrix}
\usage{
GraphToEdgeMatrix(g)
}
\arguments{
\item{g:}{a graphNEL object}
}
\value{
Edge Matrix
}
\description{
converts graphNEL to an edge matrix in the format of \code{{from, to}}
}
