% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GraphToEdgeMatrixWithAttributes}
\alias{GraphToEdgeMatrixWithAttributes}
\title{Convert graph to edge matrix with edge attributes}
\usage{
GraphToEdgeMatrixWithAttributes(g, attrs = c("capacity", "load", "weight"))
}
\arguments{
\item{g}{network graph}

\item{attrs}{vector of strings, listing the edge attributies}
}
\value{
edge matrix with added columns for edge attributes
}
\description{
Convert graph to edge matrix with edge attributes
}
