% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperCentralNode}
\alias{HelperCentralNode}
\title{calculate node centrality}
\usage{
HelperCentralNode(g)
}
\arguments{
\item{g}{network graph (igraph object)}
}
\value{
vector of cenrality for each node
}
\description{
calculate the centrality of all nodes in graph
}
