% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperConvertCharToInt}
\alias{HelperConvertCharToInt}
\title{convert graph nodes from character to integer}
\usage{
HelperConvertCharToInt(c)
}
\arguments{
\item{c}{character}
}
\value{
integer
}
\description{
convert graph nodes from character to integer
}
