% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperConvertIntToChar}
\alias{HelperConvertIntToChar}
\title{convert graph nodes from integer to character}
\usage{
HelperConvertIntToChar(i)
}
\arguments{
\item{c}{character}
}
\value{
integer
}
\description{
convert graph nodes from integer to character
}
