% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperRoundUp}
\alias{HelperRoundUp}
\title{Round up a number}
\usage{
HelperRoundUp(x, nice = c(1, 2, 4, 5, 6, 7, 8, 9, 10))
}
\description{
taken from http://stackoverflow.com/questions/6461209/how-to-round-up-to-the-nearest-10-or-100-or-x
}
