% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperRoundUpGranular}
\alias{HelperRoundUpGranular}
\title{Round up a number}
\usage{
HelperRoundUpGranular(x, nice = c(1, 2, 4, 5, 6, 7, 8, 9, 10))
}
\description{
adapted from the the function HelperRoundUp for finer granularity
}
