% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperSummarySE}
\alias{HelperSummarySE}
\title{Summarizes data}
\usage{
HelperSummarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{a data frame.}

\item{measurevar}{the name of a column that contains the variable to be summariezed}

\item{groupvars}{a vector containing names of columns that contain grouping variables}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{conf.interval}{the percent range of the confidence interval (default is 95%)
New version of length which can handle NA's: if na.rm==T, don't count them
This is does the summary; it's not easy to understand...
Rename the "mean" column
Confidence interval multiplier for standard error
Calculate t-statistic for confidence interval:
e.g., if conf.interval is .95, use .975 (above/below), and use df=N-1}
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
}
\references{
The function is taken from Rcookbook online at: \link{http://www.cookbook-r.com/Manipulating_data/Summarizing_data/}
}
