% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperUpdatePathLoadOnAm}
\alias{HelperUpdatePathLoadOnAm}
\title{Update path load}
\usage{
HelperUpdatePathLoadOnAm(r, l)
}
\arguments{
\item{r}{sequence of nodes representing a path in the network.}

\item{l}{number of subscriptions of item of node.}
}
\value{
updated global EM
}
\description{
Update the load on the edges of a path in the global AdjacencyMatrix EM
}
\note{
CAREFUL!!!, this function updates the global variable EM

DO NOT use diag()!!!!
}
