% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperUpdatePathLoadOnEM}
\alias{HelperUpdatePathLoadOnEM}
\title{Update path load}
\usage{
HelperUpdatePathLoadOnEM(peM, l)
}
\arguments{
\item{peM}{matrix of [from, to] edges of a path (branch)}

\item{l}{number of subscriptions of item of node.}
}
\value{
updated global EM
}
\description{
Update the load on the edges of a path in the global EdgeMatrix EM.
}
\note{
CAREFUL!!!, this function updates the global variable EM
}
