% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperUpdateTreeLoadOnAm}
\alias{HelperUpdateTreeLoadOnAm}
\title{Update tree load}
\usage{
HelperUpdateTreeLoadOnAm(t, l)
}
\arguments{
\item{l}{number of subscriptions of item of node.}

\item{r}{sequence of nodes representing a path in the network.}
}
\value{
updated global EM
}
\description{
Update the load on the edges of a multicast tree in the global AdjacencyMatrix EM
}
\note{
CAREFUL!!!, this function updates the global variable EM
}
