% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{HelperUpdateTreeLoadonEM}
\alias{HelperUpdateTreeLoadonEM}
\title{Update tree load}
\usage{
HelperUpdateTreeLoadonEM(t, r, l, iv)
}
\arguments{
\item{t}{tree in the form of list, each element represent a branch in the tree}

\item{l}{load}

\item{iv}{item volume}
}
\value{
matrix of edges in the tree
}
\description{
Extract the set of edges in a tree list. Common edges will be merged into one edge indicating multicast
}
