% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{IgraphToGraphNEL}
\alias{IgraphToGraphNEL}
\title{convert igraph to graphNEL}
\usage{
IgraphToGraphNEL(g, directed = 1)
}
\arguments{
\item{g}{igraph object}

\item{directed}{boolean \code{0} undirected, \code{1} directed graph}
}
\value{
graphNEL object
}
\description{
convert igraph to graphNEL
}
\author{
Mays AL-Naday
}
