% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\name{JarSubsAssignUsersForTimeSlot}
\alias{JarSubsAssignUsersForTimeSlot}
\title{Generate Poisson distributed number of users}
\usage{
JarSubsAssignUsersForTimeSlot(nodes.population, network.population, total.users)
}
\arguments{
\item{nodes.population}{a vector nodes populations within the network}

\item{network.population}{total population of all nodes in the network, i.e.sum(nodes.population)}

\item{total.users}{total number of active users in the network <= network.population}
}
\value{
numeric value selected from a Poisson distribution
}
\description{
Generate Poisson distributed subscription counts per network location provided there are total_users in total in the system
}
\author{
Janne Riihijärvi
}
