% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{NodePopularityCloseness}
\alias{NodePopularityCloseness}
\title{assing popularity to graph nodes}
\usage{
NodePopularityCloseness(g)
}
\arguments{
\item{g}{graph object of type igraph}
}
\value{
dataframe of (node, popularity)
}
\description{
assign popularity metirc to each node based on a node property, e.g. closeness of node
}
