% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubs.R
\name{OriginPublishers}
\alias{OriginPublishers}
\title{Construct list of origins}
\usage{
OriginPublishers(cc, gud, k, a, opt, e = c(), c = FALSE)
}
\arguments{
\item{cc}{content catalogue as a list objects, each in the form of <i, p, s, f, l, ..>}

\item{gud}{(g, u, d) list, where g is graphNEL of the core network and u is the population of each core node (before distributed over access nodes), d: DM Distanc Matrix}

\item{k}{number of origins to be selected}

\item{a}{minimum storage capacity that can be placed in a node}

\item{opt}{numeric vector giving option of assignment algorithm. 1 == assign k largest population (\code{AlgoLargest}), 2 == assign k nodes with highest 'closeness' (\code{AlgoClosest}), 3 == assign using 'Swing' (\code{AlgoSwing})}

\item{e}{set of nodes to be excluded from selection. Typically because these nodes are already selected as origins.}

\item{c}{create? <FALSE | TRUE>, default FALSE, decided whether or not to create the global PUBLISHERS list}
}
\value{
list of origins in the form of <p, v, i, c>, where p is the node id of the publisher, v is the total volume of all items, i is the set of items and c is the cost of placing an origin at the node. c is not exploited at the moment.
}
\description{
Select a subset of network nodes to publish all the items in a catalogue, in a CDN this equates to 'entry point' servers
}
