% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParseCatalogue}
\alias{ParseCatalogue}
\title{parse the catalogue of items from a YAML config file}
\usage{
ParseCatalogue(config.file, config.env = "default")
}
\arguments{
\item{config.file}{.yml configuration file containing a section named items with essential elements including: {size = set size, a = exponent of distribution, iv = item volume or volumes to choose from, ig = item ignorance (i.e. fraction of popularity distribution to be ignored or assumed cached)}}

\item{config.env}{configuration enviroment, default to \code{default}}
}
\value{
design parameters of the content catalogue
}
\description{
This is listed under the 'catalogue' section of the config file
}
\note{
this function is similar to \code{ParseItems}, except the latte is used to parse the catalogue section under the alternative name \code{items}
}
