% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParseConfig}
\alias{ParseConfig}
\title{Parse configuration file in YAML format}
\usage{
ParseConfig(config.file, config.active = "default", config.env = "default")
}
\arguments{
\item{config.file}{configuration file in YAML format}

\item{config.active}{configuration enviroment of the logger}

\item{config.env}{configuration enviroment inside the YAML file, default to 'default'}
}
\value{
a list of grouped configurations. Currently, the list includes configs of: \code{dir} - i.e. directories to be used in the simulation, \code{graph}, \code{catalogue} - i.e. CDN items catalogue, \code{origin} - i.e. content originators, \code{surrogate}, i.e. edge servers, \code{dns} - i.e. DNS and \code{permutation} - i.e. number of simulation rounds, and load on the network. the permutations here are multiplicative of the parameters.
generate Edge-Origin and DNS-Edge-Origin expression permutatins
get the adjacency matrix of the graph
Calculate the Distance matrix of g
}
\description{
parse all the elements in the YAML configuration file and return a list of grouped subsets of configurations
}
\details{
This function calles a number of smaller functions that looks for pre-defined groups in the yml, if found, then the elements within the group is retruned. Currently, only groups that are defined in the template \code{config.yml} are included, to include new groups, new atom parsing function need to be developed and added to this function.
}
