% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParseCores}
\alias{ParseCores}
\title{Parse number of cores in YAML configuration file}
\usage{
ParseCores(config.file, config.env = "default")
}
\arguments{
\item{config.file}{full path to a .yml config file containing modelling graph}

\item{config.env}{configuration enviroment, default to 'default'}
}
\value{
numeric number of cores
}
\description{
parse a YAML config file to extract number of CPU cores to be utilised in parallel processing
}
