% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParseDns}
\alias{ParseDns}
\title{parse modeling DNS from a YAML config file}
\usage{
ParseDns(config.file, config.env = "default")
}
\arguments{
\item{config.file}{.yml configuration file containing a section named dns}

\item{config.env}{configuration enviroment, default to 'default'}
}
\value{
dimensioning parameters of DNS in the network - this equivelent to dimensioning CDN local DNS points
}
\description{
This is grouped under the \code{dns} list
}
