% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParseGraph}
\alias{ParseGraph}
\title{parse graph configurations from a YAML config file}
\usage{
ParseGraph(config.file, config.env = "default")
}
\arguments{
\item{config.file}{full path to a .yml config file containing modelling graph}

\item{config.env}{configuration enviroment, default to 'default'}
}
\value{
graph configurations
}
\description{
parse graph configurations from a YAML config file
}
