% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParseOrigins}
\alias{ParseOrigins}
\title{parse modeling origins from a YAML config file}
\usage{
ParseOrigins(config.file, config.env = "default")
}
\arguments{
\item{config.file}{.yml configuration file containing a section named origins}

\item{config.env}{configuration enviroment, default to 'default'}
}
\value{
dimensioning parameters of origins in the network - this equivelent to dimensioning CDN entry points
}
\description{
This is grouped under the \code{origins} list
}
