% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParsePaths}
\alias{ParsePaths}
\title{parse a YAML config file to extract directory paths}
\usage{
ParsePaths(config.file, config.env = "default")
}
\arguments{
\item{config.file}{full path to a .yml config file containing a default template}

\item{config.env}{configuration enviroment, default to 'default'}
}
\value{
directory paths
}
\description{
parse a YAML config file to extract directory paths
}
