% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParsePermutation}
\alias{ParsePermutation}
\title{parse modelling permutations from a YAML config file}
\usage{
ParsePermutation(config.file, config.env = "default")
}
\arguments{
\item{config.file}{full path to a .yml config file containing modelling permutations}

\item{config.env}{configuration enviroment, default to 'default'}
}
\value{
modelling permutations
}
\description{
parse modelling permutations from a YAML config file
}
