% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ParseSurrogates}
\alias{ParseSurrogates}
\title{parse modeling surrogates from a YAML config file}
\usage{
ParseSurrogates(config.file, config.env = "default")
}
\arguments{
\item{config.file}{.yml configuration file containing a section named surrogate}

\item{config.env}{configuration enviroment, default to 'default'}
}
\value{
dimensioning parameters of surrogates in the network - this equivelent to dimensioning CDN edge servers
}
\description{
This is grouped under the \code{surrogate} list
}
