% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.R
\name{ProcIcnCapacity}
\alias{ProcIcnCapacity}
\title{Process ICN both unicast and multicast capacity results}
\usage{
ProcIcnCapacity(cfgs)
}
\arguments{
\item{cfgs}{configurations, following a YAML template in config/config.yml which have been used to simulate the network behaviour
Todo: extend the function to plot for multiple catalogues
substitue numeric identifiers of selection policies with human-readable names}
}
\description{
process stored backhaul capacity for ICN unicast and multicast, if not stored, return notification to generate or wait until generated.
}
\note{
This function currently assumes only one catalogue permutation.
}
