% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.R
\name{ProcIpCapacity}
\alias{ProcIpCapacity}
\title{Process IP backhaul capacity results}
\usage{
ProcIpCapacity(cfgs)
}
\arguments{
\item{cfgs}{configurations, following a YAML template in config/config.yml which have been used to simulate the network behaviour
Todo: extend the function to plot for multiple catalogues
construct combinations of expressions, will be used to create file names of results objects
make sure the expression name includes the simulation number of tests, not the generation number.
create results filenames (to be loaded)
construct combinations of metadata, length of this dataframe MUST equal length of the corresponding 'x.expr' dataframe
Todo: comeup with a more generic construction of meta data for any number of parameteres for origins and surrogates.
substitue numeric identifiers of selection policies with human-readable names}
}
\description{
process stored backhaul capacity for IP, if not stored, return notification to generate or wait until generated.
}
\note{
This function currently assumes only one catalogue permutation.
}
