% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.R
\name{ProcIpPathLength}
\alias{ProcIpPathLength}
\title{Process IP path length results}
\usage{
ProcIpPathLength(cfgs, chosen.cols = c("l", "e", "dopt", "eopt", "oopt", "ek",
  "ok", "dk"))
}
\arguments{
\item{cfgs}{configurations, following a YAML template in config/config.yml which have been used to simulate the network behaviour
construct combinations of expressions, will be used to create file names of results objects
make sure the expression name includes the simulation number of tests, not the generation number.
create results filenames (to be loaded)
Todo: comeup with a more generic construction of meta data for any number of parameteres for origins and surrogates.}
}
\description{
process path length and generate ECDF of path lengths of all tests for which there are stored data. If not stored, return notification to generate or wait until generated.
}
