% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.R
\name{ProcTotalStorage}
\alias{ProcTotalStorage}
\title{Process Storage capacity}
\usage{
ProcTotalStorage(cfgs, chose.cols = c("c", "nc", "ek", "eopt", "oopt", "ok"))
}
\arguments{
\item{cfgs}{configurations, following a YAML template in config/config.yml which have been used to simulate the network behaviour
construct combinations of expressions, will be used to create file names of results objects
Todo: comeup with a more generic construction of meta data for any number of parameteres for origins and surrogates.}
}
\description{
Calculate storage capacity required to cach all items of a catalogue.
}
