% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubs.R
\name{PublishersCombineCachedWithNonCached}
\alias{PublishersCombineCachedWithNonCached}
\title{Combine cached items with non cached items in one vector}
\usage{
PublishersCombineCachedWithNonCached(c, nc)
}
\arguments{
\item{c}{cached items}

\item{nc}{non cached items}
}
\value{
vector of 1s and 0s , with 1s names set to cached items and 0s names set to non cached items.
}
\description{
Combine cached items with non cached items in one vector
}
