% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{RouteIcnMulticastDijkstra}
\alias{RouteIcnMulticastDijkstra}
\title{Route ICN multicast demands with Dijkstra}
\usage{
RouteIcnMulticastDijkstra(iL, g, t, L)
}
\arguments{
\item{iL}{list of items pub/sub matches}

\item{g}{network graph list, including (at least): graph object and Edge Matrix}

\item{t}{numerical value of the catchment interval}
}
\value{
list of: \code{trees} satisfying the demand, \code{added} link capacity to accommodate the demand and \code{total} demand that was introduced.
}
\description{
Route all ICN demands according to dijkstra's algorithm. Subscribers are assumed to be quasi-synchronous, hence routed as multicast. Number of multicast groups depends on \code{t} and \code{L}
}
\note{
There is no knowledge here if multiple publishers have equal distance to multiple subscribers that only one of them will be selected.
this will come in the improved dijkstra

\code{t} must be less than or equal \code{L}
}
