% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{RouteIcnMulticastOneDemandDijkstra}
\alias{RouteIcnMulticastOneDemandDijkstra}
\title{Route ICN multicast demands with Dijkstra}
\usage{
RouteIcnMulticastOneDemandDijkstra(iL, D, R, t, L)
}
\arguments{
\item{iL}{demands of a single item in the format of a tuple <i, f, s, p, v>, i: id(or rank), p: set of publishers, s: set of subscribers, v: volume, f: probability of occurance}

\item{D}{N x N shortest distance matrix}

\item{R}{N x N dataframe of shortest routes (min hop count)}

\item{t}{aggregation (or catchment) interval - number of seconds to wait for multicast group to build up from arriving requests before responding}

\item{L}{total transmission/streaming period}
}
\value{
list of: \code{trees} satisfying the demand, \code{added} link capacity to accommodate the demand and \code{total} demand that was introduced.
update multicast load on each tree
}
\description{
Route multicast demands of a single item, update the link weights in the global EM and report admission ratio. Subscribers are assumed to be quasi-synchronous, hence routed as multicast. Number of multicast groups depends on \code{t} and \code{L}
}
