% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{RouteIcnUnicastDijkstra}
\alias{RouteIcnUnicastDijkstra}
\title{Route ICN unicast demands with Dijkstra}
\usage{
RouteIcnUnicastDijkstra(iL, g)
}
\arguments{
\item{iL}{list of Pub/Sub state for each item in catalogue}

\item{g}{network graph list, including (at least): graph object and Edge Matrix}
}
\value{
list of: \code{trees} satisfying the demand, \code{added} link capacity to accommodate the demand and \code{total} demand that was introduced.
}
\description{
Route ICN subscriptions (demands) of all items using dijkstra's algorithm for min hop-count path. subscriptions here are assumed completely desynchronous, hence routed as unicast. Each demand is a tuple of \code{<i, s, p, v>}
}
\note{
There is no knowledge here if multiple publishers have equal distance to multiple subscribers that only one of them will be selected.
this will come in the improved dijkstra
}
