% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{RouteIcnUnicastOneDemandDijkstra}
\alias{RouteIcnUnicastOneDemandDijkstra}
\title{Route ICN demands with Dijkstra}
\usage{
RouteIcnUnicastOneDemandDijkstra(iL, D, R)
}
\arguments{
\item{iL}{demands of a single item in the format of a tuple \code{<i, f, s, p, v>}, i: id(or rank), p: set of publishers, s: set of subscribers, v: volume, f: probability of occurance}

\item{D}{N x N shortest distance matrix}

\item{R}{N x N dataframe of shortest routes (min hop count)}
}
\value{
list of: \code{trees} satisfying the demand, \code{added} link capacity to accommodate the demand and \code{total} demand that was introduced.
}
\description{
Route ICN subscriptions (demand) of a single item, update the link weights in the global EM and report admission ratio. subscriptions here are assumed completely desynchronous, hence routed as unicast.
}
