% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{RouteIpUnicastDijkstra}
\alias{RouteIpUnicastDijkstra}
\title{Route all IP demands with Dijkstra's algorithm}
\usage{
RouteIpUnicastDijkstra(g, iL, ce, eo)
}
\arguments{
\item{g}{network graph in graphNEL format.}

\item{iL}{list of items pub/sub matches}

\item{ce}{client-edge server (surrogate) mapping through DNS}

\item{eo}{edge-origin mapping for retrieving non-cached items}
}
\value{
list of: \code{trees} satisfying the demand, \code{added} link capacity to accommodate the demand and \code{total} demand that was introduced.
}
\description{
Route all IP demands according to dijkstra's algorithm. Here unicast is assumed regardless of the sych state of clients
}
\note{
there is no knowledge here if multiple publishers have equal distance to multiple subscribers that only one of them will be selected. This will come in the improved dijkstra
}
