% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{RouteIpUnicastOneDemandDijkstra}
\alias{RouteIpUnicastOneDemandDijkstra}
\title{Route IP unicast demands with Dijkstra}
\usage{
RouteIpUnicastOneDemandDijkstra(iL, D, R, ce, eo)
}
\arguments{
\item{iL}{demands of a single item in the format of a tuple <i, f, s, p, v>, i: id(or rank), p: set of publishers, s: set of subscribers, v: volume, f: probability of occurance}

\item{ce}{client-edge server (surrogate) mapping through DNS}

\item{eo}{edge-origin mapping for retrieving non-cached items}

\item{g}{network graph}
}
\value{
list of: \code{trees} satisfying the demand, \code{added} link capacity to accommodate the demand and \code{total} demand that was introduced.
}
\description{
Route IP demand for a single stream, provided DNS map between clients and surrogates (edge servers), and mapping between surrogates and origin servers. This represnets conventional CDN routing towards a surrogate selected through DNS redirect
}
