% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{RouteTestTotal}
\alias{RouteTestTotal}
\title{Testing total presented demands}
\usage{
RouteTestTotal(cc.fn)
}
\arguments{
\item{cc.fn}{file name of the cc-g_xxx object to be loaded. notice this is pub/sub state not mere catalogue.}
}
\description{
testing function to see how much is the demand introduced by a stored list of Pub/Sub state.
}
