% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetAdditionalItemPublisher}
\alias{SetAdditionalItemPublisher}
\title{Append publisher to list of publishers of item}
\usage{
SetAdditionalItemPublisher(x, p, l)
}
\arguments{
\item{x}{item object}

\item{p}{publisher(s)}

\item{l}{locally available or not.}
}
\description{
Append publisher to list of publishers of item
}
