% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetAdditionalItemSubscriber}
\alias{SetAdditionalItemSubscriber}
\title{Append subscriber to set of item's subscribers}
\usage{
SetAdditionalItemSubscriber(x, v, s)
}
\arguments{
\item{x}{item object}

\item{v}{volume (number of subscriptions)}

\item{s}{subscriber}
}
\description{
append subscriber to the set of subscribers of item in list of demands to the list
}
