% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{SetCapacityValueToEdgeMatrix}
\alias{SetCapacityValueToEdgeMatrix}
\title{Set capacity on graph links}
\usage{
SetCapacityValueToEdgeMatrix(eM, min = 5, max = 10)
}
\arguments{
\item{eM}{edge matrix, possibly generated using GraphToEdgeMatrix(g), including a column with column name \code{capacity}}

\item{min}{minimum capacity value on any edge}

\item{max}{maximum capacity on any edge}
}
\value{
edge matrix with capacity set randomly between min and max
}
\description{
Assign randomly generated capacity values to the edge matrix of a graph. Notice the matrix must have a column \code{capacity} already
}
