% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetDnsMap}
\alias{SetDnsMap}
\title{Set DNS Map with clients and surrogates}
\usage{
SetDnsMap(d, n = c(), p = c())
}
\arguments{
\item{d}{DNS state}

\item{n}{set of clients nodes}

\item{p}{set of surrogaes nodes}
}
\description{
Set DNS Map with clients and surrogates
}
