% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetDnsSurrogates}
\alias{SetDnsSurrogates}
\title{Set DNS surrogates}
\usage{
SetDnsSurrogates(d, p)
}
\arguments{
\item{d}{DNS state}

\item{p}{set of nodes}
}
\value{
updated state with d$n = n
}
\description{
Set DNS surrogates
}
