% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetEvnArgs}
\alias{SetEvnArgs}
\title{Set enviroment with configuration arguments}
\usage{
SetEvnArgs(env, ...)
}
\arguments{
\item{env}{enviroment to set the arguments within}
}
\description{
Set enviroment with configuration arguments
}
