% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetItemPublisher}
\alias{SetItemPublisher}
\title{Set publisher of item}
\usage{
SetItemPublisher(x, p, l)
}
\arguments{
\item{x}{item object}

\item{p}{publisher}

\item{l}{localizing publishers (i.e. caching)}
}
\description{
Set publisher of item
}
