% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetItemPublishersList}
\alias{SetItemPublishersList}
\title{Construct list of publishers}
\usage{
SetItemPublishersList(x, v = Inf, i = c(), c = NULL)
}
\arguments{
\item{v}{both storage and bandwideth capacity of the publisher, defined as sum of the volumes of published items}

\item{i}{set of advertised items}

\item{c}{cost of the publisher. i.e. cost of setting up a publisher in the node.}

\item{node}{label of the publisher}
}
\description{
Construct list of publishers
}
