% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetItemReflector}
\alias{SetItemReflector}
\title{Set item reflector}
\usage{
SetItemReflector(i, R)
}
\arguments{
\item{i}{item list object in the format <i, f, p, s, v,...>}

\item{R}{set of reflectors}
}
\value{
updated item list
}
\description{
Item here represents a live stream item in streaming CDN
}
