% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetItemsList}
\alias{SetItemsList}
\title{Set attributes of a catalogue element (item)}
\usage{
SetItemsList(x, f, s = c(), p = c(), l = c(), v = c())
}
\arguments{
\item{x}{item id - currently the rank is also the item id}

\item{f}{(0,1) frequency or probability of occurance following a popularity distribution model}

\item{s}{subscribers or consumers of the item}

\item{p}{publishers or providers of the item}

\item{l}{boolean giving locality of the item within each publisher - i.e. 1 item is stored locally, 0 item is not stored locally}

\item{v}{volume of the item in Gbps}
}
\description{
create items list of demands in the form of (i,f, S,P) - i.e. <item, fraction of total occurances, Subscribers:#Subscriptions, Publishers>, volume is assumed 1
}
