% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetPublisherCapacity}
\alias{SetPublisherCapacity}
\title{Set publisher capacity}
\usage{
SetPublisherCapacity(p, u, v, a)
}
\arguments{
\item{p}{publisher node}

\item{u}{numeric vector of population per node in the network}

\item{v}{volume to be approtioned to the surrogate}

\item{a}{minimum volume that can be placed in a node, this is a dimensioning parameter that depends on server specs}
}
\description{
Set publisher capacity in terms of item volumes it can publish. Decision here is based on the ratio of the node population to the largest node population
}
