% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetReplacementValue}
\alias{SetReplacementValue}
\title{Substitue numeric value of column with a human-readable name}
\usage{
SetReplacementValue(df, cols, ov, nv)
}
\arguments{
\item{df}{dataframe of the cloumn}

\item{cols}{column names where the replacement will be}

\item{ov}{old value, i.e. existing value. the position of a value here MUST match the position of the name in the nv param (next)}

\item{nv}{new value, the new value to replace ov. The position of each element in nv must correspond to the same position in ov}
}
\description{
Substitue numeric value of column with a human-readable name
}
