% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{SetSubscribersList}
\alias{SetSubscribersList}
\title{Set attributes of subscribers list}
\usage{
SetSubscribersList(x, v, i)
}
\arguments{
\item{x}{node label of the subsciber}

\item{v}{number of subscriptions}

\item{i}{set of 'subscribed to' items}
}
\value{
list object of the subscriber
}
\description{
Create subscribers list of demands in the form of (s,v,i) - i.e. <subscriber, volume, set of items>
}
