% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimIcnMulticast}
\alias{SimIcnMulticast}
\title{simulate routing ICN demands in a native ICN or *-over-ICN network}
\usage{
SimIcnMulticast(cfgs)
}
\arguments{
\item{cfgs}{configurations as parsed from config file in YAML format, config/config.yml
define output containers
define prallel cluster of workers
construct permutations names and load relevant input data
set the number of tests to be simulated, if not present in cfgs$tcfg (i.e. it has not be set in config.yml)
make sure number of simulation tests is smaller or equal to the number of generation tests, otherwise, its invalide data situation
make sure the expression name includes the simulation number of tests, not the generation number.
create subdirectories for the results of each catalogue
create results directory
construct results filenames
cc[[g]] is 3-Level list:
 EO.combo
 |-- scale.index
     |-- test.number}
}
\description{
simulate routing ICN demands in a native ICN or *-over-ICN network
}
