% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimIcnMulticastInstance}
\alias{SimIcnMulticastInstance}
\title{Simulate routing ICN demands in a native ICN or *-over-ICN network}
\usage{
SimIcnMulticastInstance(cc, g, scale, tests, t, L, path)
}
\description{
simulate provisioned capacity for ICN multicast traffic, for a single catchement interval and single length of running

 @param cc pub/sub state for a single content catalogue
 @param g network graph list including graph object, per node population, name, distance matrix and edge matrix
 @param scale ratio of population as active users
 @param tests number of tests to run
 @param t catchement interval in seconds
 @param L total length in seconds
 @param path full path as to where to store the data, including the filename
}
