% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimIcnUnicastInstance}
\alias{SimIcnUnicastInstance}
\title{simulate provisioned capacity for ICN unicast traffic}
\usage{
SimIcnUnicastInstance(cc, g, scale, tests, path)
}
\description{
@param cc pub/sub state for a single content catalogue
 @param g network graph list including graph object, per node population, name, distance matrix and edge matrix
 @param scale ratio of population as active users
 @param tests number of tests to run
 @param path full path as to where to store the data, including the filename
}
