% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimIp}
\alias{SimIp}
\title{simulate routing ICN demands in a native ICN or *-over-ICN network}
\usage{
SimIp(cfgs)
}
\arguments{
\item{cfgs}{configurations as parsed from config file in YAML format, config/config.yml
define prallel cluster of workers
construct permutations names and load relevant input data
create subdirectories for the results of each catalogue
set the number of tests to be simulated, if not present in cfgs$tcfg (i.e. it has not be set in config.yml)
make sure number of simulation tests is smaller or equal to the number of generation tests, otherwise, its invalide data situation
make sure the expression name includes the simulation number of tests, not the generation number.
create results directory
construct results filenames
cc RV state only reflects the combinations of origins and surrogates, hence the list length is length(origins combo) x length(surrogate combo).
To evaluate IP allocation of the cc state for different DNS combo, each element need to be repeated length(DNS combo) times
cc[[g]] is 3-Level list:
 EO.combo
 |-- scale.index
     |-- test.number}
}
\description{
simulate routing ICN demands in a native ICN or *-over-ICN network
}
