% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimIpInstance}
\alias{SimIpInstance}
\title{simulation one instance for the permutation configuration and the load}
\usage{
SimIpInstance(cc, g, cemap, eomap, scale, tests, path)
}
\arguments{
\item{cc}{Pub/Sub states for content catalogues of specific configuration set}

\item{g}{network graph list, including: graph object, population/node, graph.name, Distance Matrix, Edge Matrix}

\item{cemap}{including \code{ce}: client-to-surrogate(edge server) mapping according to DNS redirect}

\item{eomap}{surrogate-to-origin mapping}

\item{scale}{scaling load on the network}

\item{tests}{number of permutations}

\item{path}{full path - including filename - to store the results}

\item{return}{list of provisioned capacity
define dataframes for results}
}
\description{
simulation one instance for the permutation configuration and the load
}
