% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\name{SubscribePerItemPerNode}
\alias{SubscribePerItemPerNode}
\title{Subscribe a node to set of items}
\usage{
SubscribePerItemPerNode(u, I)
}
\arguments{
\item{u}{access population, i.e. the number of active users associated with an access node}

\item{I}{a sorted set, representing the global set of items in the network. I is the probabilities of ITEMS, the global list of items.}
}
\value{
list of subscriptions per item per node
}
\description{
Generate subscriptions per item per node by approtioning the total number of subscriptions according to some probability distribution such as Zipf
}
