% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\name{SubscribePerNode}
\alias{SubscribePerNode}
\title{Subscribe every node to a set of items}
\usage{
SubscribePerNode(u, I, c = FALSE)
}
\arguments{
\item{u}{number of active users associated with the node}

\item{I}{a sorted set, representing the global set of items in the network. I is the probabilities of ITEMS, the global list of items.}

\item{c}{create?<FALSE|TRUE>, default FALSE, decided whether or not to create the global SUBSCRIBERS list}
}
\value{
list of subscribers, each of which is a tuple object (s,v,i), where s is subscriber id, v numeric vector of subscriptions per item, i is character vector of items Ids
}
\description{
Generate subscriptions per node, i.e. associate a set of items (subset of the global set of items) with a node.
}
