% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubs.R
\name{SurrogatePublishers}
\alias{SurrogatePublishers}
\title{Construct list of surrogates}
\usage{
SurrogatePublishers(cc, gu, k = NULL, a, D = NULL, opt, e = c(),
  c = FALSE)
}
\arguments{
\item{cc}{content catalogue as a list objects, each in the form of <i, p, s, f, l, ..>}

\item{gu}{(g, u) graph, population list, where g is graphNEL of the either the core, access or core/access augumented network and u is the population of each node}

\item{k}{number of surrogates to be selected}

\item{a}{minimum storage capacity that can be placed in a node}

\item{D}{matrix of the distance from each node to every other node}

\item{opt}{numeric vector giving option of assignment algorithm. 1 == assign k largest population, 2 == assign k nodes with highest 'closeness' centrality}

\item{e}{set of nodes to be excluded from selection. Typically because these nodes are already selected as origins.}

\item{c}{create?<FALSE|TRUE>, default FALSE, decided whether or not to create the global PUBLISHERS list}
}
\value{
list of surrogates in the form of <p, v, i, c>, where p is the node id of the publisher, v is the total volume of all items, i is the set of items and c is the cost of placing an origin at the node. c is not exploited at the moment.
}
\description{
Select a subset of nodes to act as surrogates, who will advertise a subset of items from the content catalogue and have local copies, while having other items advertised but not available locally.

Items that are not available locally, can be retreived from another surrogate or an origin.

Items that are available locally will appear in the set of items p$i and the surrogate will appeare in the set of providers of an item i$p with availability boolean set to \code{TRUE}.

Items that are NOT available locally will not apprear in the set of items p$i and the surrogate will be marked with availability boolean FALSE in an item object i$p
}
