% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{vGenOriginsPermutations}
\alias{vGenOriginsPermutations}
\title{vectorised wrapper of origins.permutations function}
\usage{
vGenOriginsPermutations(.opt, .k, .a, .cc, .gund, .tests, .path)
}
\description{
vectorised wrapper of origins.permutations function
}
