% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{vGenSurrogatePermutations}
\alias{vGenSurrogatePermutations}
\title{vectorised wrapper of origins.permutations function}
\usage{
vGenSurrogatePermutations(.opt, .k, .a, .cc, .gund, .e, .tests, .path)
}
\description{
vectorised wrapper of origins.permutations function
}
