% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{vGetCatalogue}
\alias{vGetCatalogue}
\title{verctorised form of \code{GetCatalogue}}
\usage{
vGetCatalogue(N, d, d.param, v, path)
}
\description{
verctorised form of \code{GetCatalogue}
}
\seealso{
\code{\link{GetCatalogue}}
}
